requirements.txt
appnope==0.1.0
arrow==0.12.1
backcall==0.1.0
boto3==1.7.26
botocore==1.10.26
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
Django==2.0.5
django-debug-toolbar==1.9.1
django-jet==1.0.7
django-solo==1.1.3
django-storages==1.6.6
django-tenant-schemas==1.9.0
django-widget-tweaks==1.4.2
docutils==0.14
flake8==3.5.0
google-api-python-client==1.4.1
greenlet==0.4.13
httplib2==0.11.3
idna==2.6
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
jmespath==0.9.3
mccabe==0.6.1
msgpack==0.5.6
neovim==0.2.6
oauth2client==4.1.2
pagseguro==0.3.2
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pyactiveresource==2.1.2
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.4
PyYAML==5.1.1
requests==2.18.4
reverse==0.1.0
rsa==3.4.2
s3transfer==0.1.13
ShopifyAPI==5.1.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
tld==0.7.10
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
xmltodict==0.11.0
